feat: Add nonCoulomb_kokkos subdirectory for Kokkos support
Add a new subdirectory called `nonCoulomb_kokkos` to the `src/potential` directory. This subdirectory is added to support Kokkos in the project. The addition of this subdirectory is conditional on the `BUILD_WITH_KOKKOS` flag, which is checked in the CMakeLists.txt file. If the flag is set, the `nonCoulomb_kokkos` subdirectory is included in the build process.