feat: Add KokkosLennardJones class and transfer methods
The code changes introduce a new class `KokkosLennardJones` in the `potential` namespace. This class is responsible for transferring data from the `matrix_shared_pair` to Kokkos data structures. The transfer method `transferFromNonCoulombPairMatrix` is added to facilitate the data transfer. This change is part of the ongoing update to support Kokkos in the project.