feat: Update KokkosSimulationBox transfer methods
The code changes in `simulationBox_kokkos.cpp` update the transfer methods in the `KokkosSimulationBox` class. The `transferAtomTypesFromSimulationBox`, `transferMolTypesFromSimulationBox`, `transferInternalGlobalVDWTypesFromSimulationBox`, `transferPositionsFromSimulationBox`, and `initializeForces` methods now use `Kokkos::deep_copy` to transfer data between host and device views. This update is part of the ongoing effort to support Kokkos in the project.