feat: Add support for MPI in simulationBox
The code changes in `simulationBox/CMakeLists.txt` add support for MPI in the `simulationBox` module. When the `BUILD_WITH_MPI` flag is enabled, the `simulationBox_MPI.cpp` file is included in the source files. This allows for parallel execution using MPI. This change is necessary to enhance the performance and scalability of the simulationBox module.