feat: Add coulomb_kokkos directory to CMakeLists.txt
The code changes in `src/potential/CMakeLists.txt` add the `coulomb_kokkos` directory to the CMakeLists.txt file. This directory is now included in the build when the `BUILD_WITH_KOKKOS` flag is enabled. This change allows for the compilation and integration of the `coulomb_kokkos` module into the project.