refactor: Add shift forces to KokkosSimulationBox
This commit adds the `_shiftForces` member variable to the `KokkosSimulationBox` class in `simulationBox_kokkos.hpp`. It also includes the `initializeShiftForces` and `transferShiftForcesToSimulationBox` methods in `simulationBox_kokkos.cpp` to initialize and transfer the shift forces between the host and device views. This addition is necessary to support the calculation of shift forces in the simulationBox module.