refactor: Update Timings class to use Duration for time calculations
This refactor updates the Timings class in timings.hpp to use the Duration type from the chrono library for time calculations. It also adds a new member variable `_timings` of type vector<TimingsManager> to store timing information. This change improves the accuracy and flexibility of time measurements in the Timings class.