feat: Add default number of epochs for optimizer
This commit adds a default number of epochs for the optimizer. The `_N_EPOCHS_DEFAULT_` constant is defined in the `defaults.hpp` file and set to 100. This default value will be used if the number of epochs is not specified in the input file. The `OptimizerSettings` class has been updated to include a static method `setNumberOfEpochs` to set the number of epochs.