refactor: Update optEngine to use std::unique_ptr for optimizer
This commit updates the optEngine.hpp and optEngine.tpp.hpp files to use std::unique_ptr for the optimizer. The _optimizer member variable is now a std::unique_ptr<optimizer::Optimizer> instead of an optimizer::Optimizer object. This change improves memory management and allows for more flexibility in choosing the optimizer type.