refactor: Add steepest descent optimizer
This commit adds the steepestDescent.hpp and steepestDescent.cpp files to the optimizer module. The SteepestDescent class is implemented as a subclass of the Optimizer class and provides functionality for performing optimization using the steepest descent method. This addition enhances the optimizer module by offering a new optimization algorithm option.