refactor: Update Optimizer class to check convergence of properties
This commit updates the Optimizer class to include a new method, hasPropertyConv(), which checks if a property of the optimizer has converged according to the convergence strategy set. It also adds a new member function, hasConverged(), which uses the hasPropertyConv() method to determine if the optimizer as a whole has converged. These changes improve the convergence checking functionality of the Optimizer class.