refactor: Update LogOutput class to write convergence warning
This commit adds a new method, writeConvergenceWarning(), to the LogOutput class. The method writes a warning message to the log file if the optimization did not converge. It takes a message parameter to provide additional information about the convergence failure. This change enhances the functionality of the LogOutput class by providing better feedback on convergence issues.