refactor: Update LearningRateStrategy constructor to accept frequency parameter
The code changes in learningRateStrategy.hpp and learningRateStrategy.cpp update the LearningRateStrategy class to include a new constructor that accepts a frequency parameter. This allows the user to specify the frequency at which the learning rate is updated. By default, the frequency is set to 1. This refactor improves the flexibility and configurability of the LearningRateStrategy class.