refactor: Add clone method to ForceField class
This refactor adds a clone method to the ForceField class in forceFieldClass.hpp and forceFieldClass.cpp files. The clone method creates a shared pointer to a new ForceField object that is a copy of the current object. This change improves the codebase by providing a convenient way to create a deep copy of ForceField objects.