refactor: Add clone method to IntraNonBonded class
This refactor adds a clone method to the IntraNonBonded class in intraNonBonded.hpp and intraNonBonded.cpp files. The clone method creates a shared pointer to a new IntraNonBonded object that is a copy of the current object. This change improves the codebase by providing a convenient way to create a deep copy of IntraNonBonded objects.