refactor: Add clone method to PhysicalData class
This refactor adds a clone method to the PhysicalData class in physicalData.hpp and physicalData.cpp files. The clone method creates a shared pointer to a new PhysicalData object that is a copy of the current object. This change improves the codebase by providing a convenient way to create a deep copy of PhysicalData objects.