refactor: Add clone method to SimulationBox class
This refactor adds a clone method to the SimulationBox class in simulationBox.hpp and simulationBox.cpp files. The clone method creates a shared pointer to a new SimulationBox object that is a copy of the current object. This change improves the codebase by providing a convenient way to create a deep copy of SimulationBox objects.