refactor: Add clone method to CellList class
This refactor adds a clone method to the CellList class in celllist.hpp and celllist.cpp files. The clone method creates a shared pointer to a new CellList object that is a copy of the current object. This change improves the codebase by providing a convenient way to create a deep copy of CellList objects.