refactor: Add clone method to MMEvaluator class
This refactor adds a clone method to the MMEvaluator class in mmEvaluator.hpp and mmEvaluator.cpp files. The clone method creates a shared pointer to a new MMEvaluator object that is a copy of the current object. This change improves the codebase by providing a convenient way to create a deep copy of MMEvaluator objects.