refactor: Add norms function for std::vector of Vector3D objects
The commit adds a new function `norms` to the `linearAlgebra` namespace. This function calculates the norms of a `std::vector` of `Vector3D` objects and returns a `std::vector` of the same type. The function iterates over each element in the input vector and calls the `norm` function to calculate the norm of each `Vector3D` object. The calculated norms are then stored in a new `std::vector` and returned.

This commit modifies the following files:
- `vector3d.hpp`
- `vector3d.tpp.hpp`