refactor: Add max and maxNorm functions for std::vector of Vector3D objects
The commit adds new functions `max` and `maxNorm` to the `linearAlgebra` namespace. The `max` function calculates the maximum value among the maximums of all Vector3D objects in a std::vector. The `maxNorm` function calculates the maximum of the norms of all Vector3D objects in a std::vector. These functions are implemented using the `std::ranges::max_element` algorithm from the `<algorithm>` header.

This commit modifies the following files:
- `vector3d.hpp`
- `vector3d.tpp.hpp`