refactor: Add getRMSForce and getMaxForce functions to Optimizer class
The commit adds new functions `getRMSForce` and `getMaxForce` to the `Optimizer` class in the `opt` namespace. These functions retrieve the root mean square (RMS) force and the maximum force, respectively, from the history of forces stored in the `Optimizer` object. The functions are implemented using the `getRMSForce` and `getMaxForce` methods of the `SimulationBox` class.

This commit modifies the following files:
- `include/opt/optimizer/optimizer.hpp`
- `src/opt/optimizer/optimizer.cpp`
- `src/engine/optEngine.cpp`