refactor: Add parseOptFilename function to inputFileParserOutput
This commit adds the `parseOptFilename` function to the `inputFileParserOutput` class in the `input` namespace. The function is responsible for parsing the optimization filename of a simulation and adding it to the output. The default value for the optimization filename is `default.opt`. The function is implemented in the `inputFileParserOutput.cpp` file.

Files modified:
- `include/input/inputFileParser/inputFileParserOutput.hpp`
- `src/input/inputFileParser/inputFileParserOutput.cpp`