refactor: Update QMSettings to include MaceModelSize and MaceModelPath
This commit updates the QMSettings class to include two new member variables: _maceModelSize and _maceModelPath. It also adds corresponding setter and getter methods for these variables. The purpose of these changes is to enhance the flexibility and configurability of the QMSettings class by allowing the user to specify the size and path of the Mace model to be used in the QM calculations.