refactor: Update QMSettings to include MaceModelType
This commit updates the QMSettings class to include a new member variable: _maceModelType. It also adds corresponding setter and getter methods for this variable. The purpose of these changes is to enhance the flexibility and configurability of the QMSettings class by allowing the user to specify the type of the Mace model to be used in the QM calculations.