refactor: Parse and handle Mace QM method in QMParser
This commit adds support for parsing and handling the Mace QM method in the QMParser class. It introduces a new method, parseMaceQMMethod, which is called when the QM method starts with "mace". The method sets the appropriate Mace model type based on the input and updates the QMSettings accordingly. This change improves the flexibility and configurability of the QMParser by allowing the user to specify different Mace models for QM calculations.