refactor: Set Mace QMRunner in QMMDEngine
This commit adds a new method, setMaceQMRunner, to the QMMDEngine class. The method sets the QMRunner object for Mace type QM methods. It checks if ASE was enabled at compile time and throws a CompileTimeException if it was not. This change improves the configurability of the QMMDEngine by allowing the user to specify different QMRunners for Mace type QM calculations.