refactor: started to refactor qm run function and potential calculate
For QM run() and MM calculation() functions new additional run/calculation() functions are implemented respectivley. These functions have a vector of molecules as additional parameter and use a selectPartitionBox method of the SimulationBox class to choose a partition of the Simulation Box. No new tests are implemented until yet. The selectPartitionBox method have to adapted regarding the handlingen of qmAtoms and Atoms.