fix: insert qm energy into force file when pc are present
In the dftbplus_periodic_stress script the FORCE file
has to start with the total energy from the QM calculation.
When point charges are present, the script runs dftb+ twice,
once with and once without point charges. The energy from the
first run is stored in a variable named $qm_energy, which is
printed to the FORCE file at the beginning, ensuring the correct format.
Before this change, the force array had an off-by-one error, when
point charges were used.